fix: address review feedback for anidb warning modal
Extracts the AniDB warning modal into dedicated files, hardens the mass-add trigger against bypassing the warning flow, and aligns browser/date utility usage with project conventions.
